Removed vue-app and moved all logic to react-app
