--- examples/vue-app/public/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width,initial-scale=1.0" />
-    <link rel="icon" href="<%= BASE_URL %>favicon.ico" />
-    <title>json-as-xlsx</title>
-  </head>
-  <body>
-    <noscript>
-      <strong>We're sorry but this app doesn't work without JavaScript</strong>
-      <strong>Please enable it to continue</strong>
-    </noscript>
-    <div id="app"></div>
-  </body>
-</html>
